LayoutTests/platform/chromium/fast/events/touch/gesture/pad-gesture-fling-expected.txt

Tests basic use of GestureFlingStart

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clientX is 10
PASS event.clientY is 11
FAIL event.wheelDeltaX > 5 should be true. Was false.
FAIL event.wheelDeltaY > 5 should be true. Was false.
PASS event.clientX is 10
PASS event.clientY is 11
FAIL event.wheelDeltaX > 5 should be true. Was false.
FAIL event.wheelDeltaY > 5 should be true. Was false.
PASS actualWheelEventsOccurred is 2
PASS successfullyParsed is true

TEST COMPLETE
